Promote Tockilator::err.
I noticed that err doesn't actually use its &self parameter --
unnecessary self-borrows are a thing that can tie you into knots, so I
removed it.

I then noticed that it's a general pattern that isn't specific to the
Tockilator type, so I made it into a normal fn and simplified some
parsing code.